+++ internal/topic/topicwriterinternal/writer_reconnector.go
@@ -325,6 +325,10 @@ func (w *WriterReconnector) createMessagesWithContent(messages []PublicMessage) 
 	return res, nil
 }
 
+func (w *WriterReconnector) Flush(ctx context.Context) error {
+	return w.queue.WaitLastWritten(ctx)
+}
+
 func (w *WriterReconnector) Close(ctx context.Context) error {
 	return w.close(ctx, xerrors.WithStackTrace(errStopWriterReconnector))
 }