+++ internal/topic/topicwriterinternal/writer_reconnector_test.go
@@ -530,6 +530,70 @@ func TestWriterImpl_Reconnect(t *testing.T) {
 	})
 }
 
+func TestWriterImpl_CloseWithFlush(t *testing.T) {
+	e := newTestEnv(t, nil)
+
+	messageTime := time.Date(2023, 9, 7, 11, 34, 0, 0, time.UTC)
+	messageData := []byte("123")
+
+	const seqNo = 36
+
+	writeCompleted := make(empty.Chan)
+	e.stream.EXPECT().Send(&rawtopicwriter.WriteRequest{
+		Messages: []rawtopicwriter.MessageData{
+			{
+				SeqNo:            seqNo,
+				CreatedAt:        messageTime,
+				UncompressedSize: int64(len(messageData)),
+				Partitioning:     rawtopicwriter.Partitioning{},
+				Data:             messageData,
+			},
+		},
+		Codec: rawtopiccommon.CodecRaw,
+	}).Return(nil)
+
+	closeCompleted := make(empty.Chan)
+	go func() {
+		err := e.writer.Write(e.ctx, []PublicMessage{{
+			SeqNo:     seqNo,
+			CreatedAt: messageTime,
+			Data:      bytes.NewReader(messageData),
+		}})
+		close(writeCompleted)
+		require.NoError(t, err)
+	}()
+
+	<-writeCompleted
+
+	go func() {
+		require.NoError(t, e.writer.Flush(e.ctx))
+		require.NoError(t, e.writer.Close(e.ctx))
+		close(closeCompleted)
+	}()
+
+	select {
+	case <-closeCompleted:
+		t.Fatal("flush and close must complete only after message is acked")
+	case <-time.After(100 * time.Millisecond):
+		// pass
+	}
+
+	e.sendFromServer(&rawtopicwriter.WriteResult{
+		Acks: []rawtopicwriter.WriteAck{
+			{
+				SeqNo: seqNo,
+				MessageWriteStatus: rawtopicwriter.MessageWriteStatus{
+					Type:          rawtopicwriter.WriteStatusTypeWritten,
+					WrittenOffset: 4,
+				},
+			},
+		},
+		PartitionID: e.partitionID,
+	})
+
+	xtest.WaitChannelClosed(t, closeCompleted)
+}
+
 func TestAllMessagesHasSameBufCodec(t *testing.T) {
 	t.Run("Empty", func(t *testing.T) {
 		require.True(t, allMessagesHasSameBufCodec(nil))
@@ -891,7 +955,7 @@ func newTestEnv(t testing.TB, options *testEnvOptions) *testEnv {
 	require.NoError(t, res.writer.waitFirstInitResponse(res.ctx))
 
 	t.Cleanup(func() {
-		res.writer.close(context.Background(), errors.New("stop writer test environment"))
+		_ = res.writer.close(context.Background(), errors.New("stop writer test environment"))
 		close(res.stopReadEvents)
 		<-streamClosed
 	})